Trying to build a radio chat system using E31ttl
This project is mainly a C++ library to manage e31ttl Transceiver in order to use it as a radio controller but for now I'm using it to implement a chat infrastructure. The main windows class is very messy, I'm just using it to work on two trancseivers at the same time, testing and debugging the program and has no other use. the setting section is also incomplete.